[BE] feat: signaling server issue 해결 (#30)
